Use hpp extension for headers (#157)
