+++ include/web_video_server/image_streamer.hpp
@@ -5,7 +5,7 @@
 #include <image_transport/image_transport.hpp>
 #include <image_transport/transport_hints.hpp>
 #include <opencv2/opencv.hpp>
-#include "web_video_server/utils.h"
+#include "web_video_server/utils.hpp"
 #include "async_web_server_cpp/http_server.hpp"
 #include "async_web_server_cpp/http_request.hpp"
