+++ include/web_video_server/ros_compressed_streamer.hpp
@@ -2,10 +2,10 @@
 #define ROS_COMPRESSED_STREAMERS_H_
 
 #include <sensor_msgs/msg/compressed_image.hpp>
-#include "web_video_server/image_streamer.h"
+#include "web_video_server/image_streamer.hpp"
 #include "async_web_server_cpp/http_request.hpp"
 #include "async_web_server_cpp/http_connection.hpp"
-#include "web_video_server/multipart_stream.h"
+#include "web_video_server/multipart_stream.hpp"
 
 namespace web_video_server
 {
