+++ include/web_video_server/vp8_streamer.hpp
@@ -38,7 +38,7 @@
 #define VP8_STREAMERS_H_
 
 #include <image_transport/image_transport.hpp>
-#include "web_video_server/libav_streamer.h"
+#include "web_video_server/libav_streamer.hpp"
 #include "async_web_server_cpp/http_request.hpp"
 #include "async_web_server_cpp/http_connection.hpp"
