+++ include/web_video_server/vp9_streamer.hpp
@@ -2,7 +2,7 @@
 #define VP9_STREAMERS_H_
 
 #include <image_transport/image_transport.hpp>
-#include "web_video_server/libav_streamer.h"
+#include "web_video_server/libav_streamer.hpp"
 #include "async_web_server_cpp/http_request.hpp"
 #include "async_web_server_cpp/http_connection.hpp"
