+++ include/web_video_server/web_video_server.hpp
@@ -10,7 +10,7 @@
 #endif
 
 #include <vector>
-#include "web_video_server/image_streamer.h"
+#include "web_video_server/image_streamer.hpp"
 #include "async_web_server_cpp/http_server.hpp"
 #include "async_web_server_cpp/http_request.hpp"
 #include "async_web_server_cpp/http_connection.hpp"
