+++ src/jpeg_streamers.cpp
@@ -1,4 +1,4 @@
-#include "web_video_server/jpeg_streamers.h"
+#include "web_video_server/jpeg_streamers.hpp"
 #include "async_web_server_cpp/http_reply.hpp"
 
 namespace web_video_server
