+++ src/multipart_stream.cpp
@@ -1,4 +1,4 @@
-#include "web_video_server/multipart_stream.h"
+#include "web_video_server/multipart_stream.hpp"
 #include "async_web_server_cpp/http_reply.hpp"
 
 namespace web_video_server
