+++ src/utils.cpp
@@ -1,4 +1,4 @@
-#include "web_video_server/utils.h"
+#include "web_video_server/utils.hpp"
 
 namespace web_video_server
 {
