+++ src/vp9_streamer.cpp
@@ -1,4 +1,4 @@
-#include "web_video_server/vp9_streamer.h"
+#include "web_video_server/vp9_streamer.hpp"
 
 namespace web_video_server
 {
